requirements.txt
Adafruit_Blinka==8.40.1
adafruit_circuitpython_ahtx0==1.0.21
matplotlib==3.8.0
matplotlib==3.8.2
numpy==1.26.4
paho_mqtt==2.1.0
pandas==2.2.2
PyQt5==5.15.10
PyQt5_sip==12.13.0
